Prevent unsafe narrowing: base/threading
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3709036
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1016422}

+++ base/threading/scoped_blocking_call_internal.cc
@@ -340,7 +340,8 @@ UncheckedScopedBlockingCall::UncheckedScopedBlockingCall(
     // Also record the data for extended crash reporting.
     const TimeTicks now = TimeTicks::Now();
     auto& user_data = scoped_activity_.user_data();
-    user_data.SetUint("timestamp_us", now.since_origin().InMicroseconds());
+    user_data.SetUint("timestamp_us", static_cast<uint64_t>(
+                                          now.since_origin().InMicroseconds()));
     user_data.SetUint("blocking_type", static_cast<uint64_t>(blocking_type));
   }
 }
